chore: rename MainAccordion to MainSection
